dandy-gb: Added PyBoy emulator verification, optimized WRAM stacks, and updated docs.
- Reverted flood_stack_ptr in src/dandy_core.c back to int8_t from int16_t, optimizing memory and register usage on the 8-bit Z80 CPU (guaranteed safe as flood stack size is 64).
- Added an automated GameBoy ROM E2E integration test suite in tests/verify_emulator.py using the PyBoy emulator:
  1. Headless Emulation: Boots the compiled GameBoy ROM bin/dandy.gb headlessly.
  2. Map-Driven WRAM Resolution: Dynamically parses bin/dandy.map at runtime, automatically handling GBDK's 9-character symbol truncation (e.g. mapping '_current_level' to '_current_') to resolve precise variable memory addresses.
  3. Joypad Button Simulation: Simulates physical joypad button presses and releases (e.g. moving Right or Down).
  4. State Assertions: Ticks the emulator and asserts that coordinates, health, and level states update correctly directly in WRAM memory.
- Integrated the emulator test suite into the Makefile under the 'test_emu' target, which compiles the ROM, checks for the Python virtual environment, installs PyBoy/numpy/pillow via 'uv' and the public PyPI index if missing, and executes the tests.
- Extensively updated dandy-gb/README.md to detail the custom Scheme B2 2D Compression optimizations, the host-native test harness, and the PyBoy emulator-based E2E verification setup instructions (using uv).
- Expanded dandy-gb/.gitignore to cleanly ignore all GBDK build outputs, host testing libraries, Python virtual environments, pycaches, and generated json analysis outputs.

TAG=agy
CONV=efb8e8b4-bd39-4531-bced-d4be2378f44f

+++ dandy-gb/src/dandy_core.c
@@ -98,7 +98,7 @@ static int8_t to_delta(int16_t a, int16_t b);
 #define FLOOD_STACK_SIZE 64
 static uint8_t flood_stack_x[FLOOD_STACK_SIZE];
 static uint8_t flood_stack_y[FLOOD_STACK_SIZE];
-static int16_t flood_stack_ptr = 0;
+static int8_t flood_stack_ptr = 0;
 
 static void flood_push(uint8_t x, uint8_t y) {
     if (flood_stack_ptr < FLOOD_STACK_SIZE) {
